requirements-dev.txt
django-querycount==0.7
django-debug-toolbar==3.5
nplusone==1
django_migration_linter==4.1.0
django_test_migrations==1.2
django-extra-checks==0.11
flake8==4.0.1
black==22.6.0
isort==5.10.1
pre-commit==2.19
pytest==7.1.2
pytest-datafiles==2.0.1
pymorphy2==0.9.1
